+++ CHANGELOG.md
@@ -40,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - URL parser incorrectly handling Markdown URLs and angled brackets
 - Intermediate bytes of CSI sequences not checked
 - Wayland clipboard integration
+- Use text mouse cursor when mouse mode is temporarily disabled with shift
 
 ## 0.3.3